Move magic commands from kernel to Jupyter project.
