+++ src/Jupyter/Magic/Simulate.cs
@@ -1,63 +1,50 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
+﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-#nullable enable
-
 using System;
-using System.Text.Json;
+using System.Diagnostics;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Jupyter.Core;
-using Microsoft.Quantum.IQSharp;
 using Microsoft.Quantum.IQSharp.Common;
-using Microsoft.Quantum.IQSharp.Jupyter;
-using Microsoft.Quantum.IQSharp.Kernel;
 using Microsoft.Quantum.Simulation.Core;
 using Microsoft.Quantum.Simulation.Simulators;
 
-namespace Microsoft.Quantum.Experimental
+namespace Microsoft.Quantum.IQSharp.Jupyter
 {
-    public class SimulateNoiseMagic : AbstractMagic
+    /// <summary>
+    ///     A magic command that can be used to simulate operations and functions
+    ///     on a full-state quantum simulator.
+    /// </summary>
+    public class SimulateMagic : AbstractMagic
     {
         private const string ParameterNameOperationName = "__operationName__";
-        private ILogger<SimulateNoiseMagic>? Logger = null;
-        private readonly INoiseModelSource NoiseModelSource;
+        private readonly IPerformanceMonitor Monitor;
 
         /// <summary>
         ///     Constructs a new magic command given a resolver used to find
         ///     operations and functions, and a configuration source used to set
         ///     configuration options.
         /// </summary>
-        public SimulateNoiseMagic(IExecutionEngine engine, ISymbolResolver resolver, IConfigurationSource configurationSource, INoiseModelSource noiseModelSource, ILogger<SimulateNoiseMagic> logger) : base(
-            "experimental.simulate_noise",
+        public SimulateMagic(ISymbolResolver resolver, IConfigurationSource configurationSource, IPerformanceMonitor monitor, ILogger<SimulateMagic> logger) : base(
+            "simulate",
             new Microsoft.Jupyter.Core.Documentation
             {
-                Summary = "Runs a given function or operation on the OpenSystemsSimulator target machine.",
-                Description = $@"
-                    > **⚠ WARNING:** This magic command is **experimental**,
-                    > is not supported, and may be removed from future versions without notice.
+                Summary = "Runs a given function or operation on the QuantumSimulator target machine.",
+                Description = @"
+                    This magic command allows executing a given function or operation on the QuantumSimulator, 
+                    which performs a full-state simulation of the given function or operation
+                    and prints the resulting return value.
 
-                    This magic command allows executing a given function or operation
-                    on the OpenSystemsSimulator target, simulating how that function or operation
-                    will perform when run on noisy quantum hardware.
-
-                    #### See also
-
-                    - [`%config`]({KnownUris.ReferenceForMagicCommand("config")})
-                    - [`%experimental.noise_model`]({KnownUris.ReferenceForMagicCommand("experimental.noise_model")})
+                    See the [QuantumSimulator user guide](https://docs.microsoft.com/azure/quantum/user-guide/machines/full-state-simulator) to learn more.
 
                     #### Required parameters
 
                     - Q# operation or function name. This must be the first parameter, and must be a valid Q# operation
                     or function name that has been defined either in the notebook or in a Q# file in the same folder.
                     - Arguments for the Q# operation or function must also be specified as `key=value` pairs.
-
-                    #### Remarks
-
-                    The behavior of this magic command can be controlled through the `%experimental.noise_model` magic command,
-                    and the `opensim.nQubits` and `opensim.representation` configuration settings.
                 ".Dedent(),
-                Examples = new string[]
+                Examples = new []
                 {
                     @"
                         Simulate a Q# operation defined as `operation MyOperation() : Result`:
@@ -74,18 +61,11 @@ namespace Microsoft.Quantum.Experimental
                         ```
                     ".Dedent(),
                 }
-            })
+            }, logger)
         {
             this.SymbolResolver = resolver;
             this.ConfigurationSource = configurationSource;
-            this.Logger = logger;
-            this.NoiseModelSource = noiseModelSource;
-
-            if (engine is IQSharpEngine iQSharpEngine)
-            {
-                iQSharpEngine.RegisterDisplayEncoder(new MixedStateToHtmlDisplayEncoder());
-                iQSharpEngine.RegisterDisplayEncoder(new StabilizerStateToHtmlDisplayEncoder(configurationSource));
-            }
+            this.Monitor = monitor;
         }
 
         /// <summary>
@@ -113,25 +93,29 @@ namespace Microsoft.Quantum.Experimental
             var inputParameters = ParseInputParameters(input, firstParameterInferredName: ParameterNameOperationName);
 
             var name = inputParameters.DecodeParameter<string>(ParameterNameOperationName);
-            var symbol = SymbolResolver.Resolve(name) as dynamic; // FIXME: Should be IQSharpSymbol.
+            var symbol = SymbolResolver.Resolve(name) as IQSharpSymbol;
             if (symbol == null) throw new InvalidOperationException($"Invalid operation name: {name}");
 
-            var qsim = new OpenSystemsSimulator(
-                ConfigurationSource.ExperimentalSimulatorCapacity,
-                ConfigurationSource.ExperimentalSimulatorRepresentation
-            );
-            if (NoiseModelSource.NoiseModel != null)
-            {
-                var json = JsonSerializer.Serialize(NoiseModelSource.NoiseModel);
-                qsim.NoiseModel = NoiseModelSource.NoiseModel;
-            }
-            Logger?.LogDebug("Simulating with noise model: {NoiseModel}", JsonSerializer.Serialize(NoiseModelSource.NoiseModel));
-            qsim.DisableLogToConsole();
-            qsim.OnLog += channel.Stdout;
+            var maxNQubits = 0L;
+
+            using var qsim = new QuantumSimulator()
+                .WithJupyterDisplay(channel, ConfigurationSource)
+                .WithStackTraceDisplay(channel);
             qsim.OnDisplayableDiagnostic += channel.Display;
-            var operation = symbol.Operation as OperationInfo;
-            var value = await operation.RunAsync(qsim, inputParameters);
-            return value.ToExecutionResult();
+            qsim.AfterAllocateQubits += (args) =>
+            {
+                maxNQubits = System.Math.Max(qsim.QubitManager.AllocatedQubitsCount, maxNQubits);
+            };
+            var stopwatch = Stopwatch.StartNew();
+            var value = await symbol.Operation.RunAsync(qsim, inputParameters);
+            stopwatch.Stop();
+            var result = value.ToExecutionResult();
+            (Monitor as PerformanceMonitor)?.ReportSimulatorPerformance(new SimulatorPerformanceArgs(
+                simulatorName: qsim.GetType().FullName,
+                nQubits: (int)maxNQubits,
+                duration: stopwatch.Elapsed
+            ));
+            return result;
         }
     }
 }
