+++ src/Kernel/Extensions.cs
@@ -17,8 +17,6 @@ using Microsoft.Jupyter.Core;
 using Microsoft.Jupyter.Core.Protocol;
 using Microsoft.Quantum.Experimental;
 using Microsoft.Quantum.IQSharp.Jupyter;
-using Microsoft.Quantum.QsCompiler.SyntaxTokens;
-using Microsoft.Quantum.QsCompiler.SyntaxTree;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
@@ -36,8 +34,8 @@ namespace Microsoft.Quantum.IQSharp.Kernel
         public static T AddIQSharpKernel<T>(this T services)
         where T: IServiceCollection
         {
-            services.AddSingleton<ISymbolResolver, Kernel.SymbolResolver>();
-            services.AddSingleton<IMagicSymbolResolver, Kernel.MagicSymbolResolver>();
+            services.AddSingleton<ISymbolResolver, SymbolResolver>();
+            services.AddSingleton<IMagicSymbolResolver, MagicSymbolResolver>();
             services.AddSingleton<IExecutionEngine, Kernel.IQSharpEngine>();
             services.AddSingleton<IConfigurationSource, ConfigurationSource>();
             services.AddSingleton<INoiseModelSource, NoiseModelSource>();
@@ -100,70 +98,6 @@ namespace Microsoft.Quantum.IQSharp.Kernel
             return version;
         }
         
-        internal static IEnumerable<QsDeclarationAttribute> GetAttributesByName(
-            this OperationInfo operation, string attributeName,
-            string namespaceName = "Microsoft.Quantum.Documentation"
-        ) =>
-            operation.Header.Attributes.Where(
-                attribute =>
-                    // Since QsNullable<UserDefinedType>.Item can be null,
-                    // we use a pattern match here to make sure that we have
-                    // an actual UDT to compare against.
-                    attribute.TypeId.Item is UserDefinedType udt &&
-                    udt.Namespace == namespaceName &&
-                    udt.Name == attributeName
-            );
-
-        internal static bool TryAsStringLiteral(this TypedExpression expression, [NotNullWhen(true)] out string? value)
-        {
-            if (expression.Expression is QsExpressionKind<TypedExpression, Identifier, ResolvedType>.StringLiteral literal)
-            {
-                value = literal.Item1;
-                return true;
-            }
-            else
-            {
-                value = null;
-                return false;
-            }
-        }
-
-        internal static IEnumerable<string> GetStringAttributes(
-            this OperationInfo operation, string attributeName,
-            string namespaceName = "Microsoft.Quantum.Documentation"
-        ) => operation
-            .GetAttributesByName(attributeName, namespaceName)
-            .Select(
-                attribute =>
-                    attribute.Argument.TryAsStringLiteral(out var value)
-                    ? value : null
-            )
-            .Where(value => value != null)
-            // The Where above ensures that all elements are non-nullable,
-            // but the C# compiler doesn't quite figure that out, so we
-            // need to help it with a no-op that uses the null-forgiving
-            // operator.
-            .Select(value => value!);
-
-        internal static IDictionary<string?, string?> GetDictionaryAttributes(
-            this OperationInfo operation, string attributeName,
-            string namespaceName = "Microsoft.Quantum.Documentation"
-        ) => operation
-            .GetAttributesByName(attributeName, namespaceName)
-            .SelectMany(
-                attribute => attribute.Argument.Expression switch
-                {
-                    QsExpressionKind<TypedExpression, Identifier, ResolvedType>.ValueTuple tuple =>
-                        tuple.Item.Length != 2
-                        ? throw new System.Exception("Expected attribute to be a tuple of two strings.")
-                        : ImmutableList.Create((tuple.Item[0], tuple.Item[1])),
-                    _ => ImmutableList<(TypedExpression, TypedExpression)>.Empty
-                }
-            )
-            .ToDictionary(
-                attribute => attribute.Item1.TryAsStringLiteral(out var value) ? value : null,
-                attribute => attribute.Item2.TryAsStringLiteral(out var value) ? value : null
-            );
 
         internal static Task<T> GetRequiredServiceInBackground<T>(this IServiceProvider services, ILogger? logger = null)
         {
@@ -200,6 +134,14 @@ namespace Microsoft.Quantum.IQSharp.Kernel
             }
 
             return dist[s1.Length, s2.Length];
+        }
+
+        internal static IServiceProvider AddBuiltInMagicSymbols(this IServiceProvider serviceProvider)
+        {
+            serviceProvider.GetRequiredService<IMagicSymbolResolver>()
+                .AddKernelAssembly<IQSharpKernelApp>()
+                .AddKernelAssembly<AzureClient.AzureClient>();
+            return serviceProvider;
         }
     }
