+++ src/Kernel/SymbolEncoders.cs
@@ -5,6 +5,7 @@ using System.Collections.Immutable;
 using System.Collections.Generic;
 using Microsoft.Jupyter.Core;
 using Markdig;
+using Microsoft.Quantum.IQSharp.Jupyter;
 
 namespace Microsoft.Quantum.IQSharp.Kernel
 {
